backend/app/requirements.txt
google-adk>=1.18.0
google-cloud-aiplatform>=1.125.0
python-dotenv>=1.0.0